EES-2793 Move the body of the filter and indicator sequence replacement to a new ReplacementServiceHelper. Also move the unit tests while leaving a basic test of each in place for the service.
